requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
reportlab
dj-database-url
django-cors-headers